% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_hdf.R
\name{build_hdf}
\alias{build_hdf}
\title{brmsMethod: build_hdf
Build a dataframe to load into the hypothesis_df function to evaluate a non-linear hypothesis using the Savage-Dickey method.}
\usage{
build_hdf(vars, priors, names, class = "b")
}
\arguments{
\item{vars}{A list containing vectors of data to be used for evaluation of non-linear hypothesis test.}

\item{priors}{A list containing vectors, each of which contain prior samples for variables loaded in vars.}

\item{names}{A vector of names corresponding to variables loaded in vars.}

\item{class}{The class of predictors with which the hypothesis will evaluate.}
}
\description{
brmsMethod: build_hdf
Build a dataframe to load into the hypothesis_df function to evaluate a non-linear hypothesis using the Savage-Dickey method.
}
