% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence_ratio.R
\name{evidence_ratio}
\alias{evidence_ratio}
\title{brmsMethod: evidence_ratio}
\usage{
evidence_ratio(
  x,
  cut = 0,
  wsign = c("equal", "less", "greater"),
  prior_samples = NULL,
  ...
)
}
\description{
brmsMethod: evidence_ratio
}
